# api/openapi.yaml
openapi: 3.0.3
info:
  title: PDn-control API
  description: API for website compliance checks, user sessions, and guest request limits.
  version: 1.1.0
servers:
  - url: http://localhost/api
    description: Docker Compose frontend proxy
paths:
  /auth/register:
    post:
      summary: Register a user and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Invalid email or weak password
        '409':
          description: Email already exists

  /auth/login:
    post:
      summary: Log in and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Invalid email or password

  /auth/logout:
    post:
      summary: Delete the current session cookie
      responses:
        '200':
          description: Logged out

  /auth/change-password:
    post:
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Weak password or invalid payload
        '401':
          description: Unauthorized or invalid current password

  /auth/me:
    get:
      summary: Get current user and guest limit state
      responses:
        '200':
          description: Current auth state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'

  /check:
    post:
      summary: Initiate a compliance check task
      description: Authenticated users are unlimited; guests are limited to 3 accepted checks per guest cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckRequest'
      responses:
        '200':
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '400':
          description: Invalid URL or check type
        '403':
          description: Guest limit reached
        '503':
          description: No workers available

  /progress/{reqId}:
    get:
      summary: Get task progress
      parameters:
        - name: reqId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskState'
        '404':
          description: Task not found

  /progress:
    post:
      summary: Internal worker progress callback
      description: Requires the X-Worker-Secret header configured on the backend.
      responses:
        '200':
          description: Progress updated
        '403':
          description: Invalid worker secret

  /reports:
    get:
      summary: Get check history for the current user
      responses:
        '200':
          description: Check history ordered newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckHistory'
        '401':
          description: Unauthorized

  /reports/{reportId}:
    get:
      summary: Download a PDF report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Report not found

  /guest/remaining:
    get:
      summary: Get remaining guest checks for the current client
      responses:
        '200':
          description: Guest limit state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestInfo'

  /workers:
    get:
      summary: Get worker pool status
      responses:
        '200':
          description: Worker status

  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy

components:
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: strong-password

    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 8
        newPassword:
          type: string
          minLength: 8

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    GuestInfo:
      type: object
      properties:
        limit:
          type: integer
          example: 3
        used:
          type: integer
          example: 1
        remaining:
          type: integer
          example: 2

    MeResponse:
      type: object
      properties:
        user:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/AuthUser'
        guest:
          $ref: '#/components/schemas/GuestInfo'

    CheckRequest:
      type: object
      required: [url, type]
      properties:
        url:
          type: string
          format: uri
          example: https://example.com
        type:
          type: string
          enum: [fast, detail]
          example: detail

    CheckResponse:
      type: object
      properties:
        code:
          type: string
          example: ERR_OK
        req-id:
          type: string
        data:
          type: object
        msg:
          type: string

    TaskState:
      type: object
      properties:
        req-id:
          type: string
        url:
          type: string
        type:
          type: string
        status:
          type: string
        worker:
          type: string
        progress:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/CheckResult'
        errors:
          type: array
          items:
            type: string
        report_id:
          type: string
        created_at:
          type: string
          format: date-time

    CheckResult:
      type: object
      properties:
        id:
          type: string
        result:
          type: string
        pages:
          type: array
          items:
            type: string
        about:
          type: string
        data:
          type: object
          additionalProperties: true

    CheckHistory:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        req_id:
          type: string
        url:
          type: string
        check_type:
          type: string
        status:
          type: string
        report_id:
          type: string
        created_at:
          type: string
          format: date-time
        file_name:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/CheckResult'